v1.0
Initial version of one pass AESVM.

Derives solution with just one pass of data. Reads data in batches 
that can fit in RAM and derive their representative set (RS) using
heuristics. The RS for each batch is written to the HDD and the 
RAM is cleared before start of read of next batch.